move util fn over
